+++ NSudo/stdafx.h
@@ -7,8 +7,6 @@
 
 // 为编译通过而禁用的警告
 #if _MSC_VER >= 1200
-// 为了不去掉未使用的形参，于是禁用该警告
-#pragma warning(disable:4100) // 未使用的形参(等级 4)
 // 编译器优化可能出现的警告（去除未引用函数并适当对一些函数使用内联）
 #pragma warning(disable:4505) // 未引用的本地函数已移除(等级 4)
 #pragma warning(disable:4710) // 函数未内联(等级 4)
@@ -28,7 +26,7 @@
 #include <Windows.h>
 
 // NSudoSDK
-#include "M2.BaseLib.hpp"
+#include "M2.NSudo.hpp"
 
 #include <commctrl.h>
 #pragma comment(lib, "comctl32.lib")